+++ lib/entities/config.dart
@@ -8,10 +8,17 @@ class Config {
   final String? packageName;
   final ArgResults? command;
 
+  // If should include untested files to report
+  final bool includeUntestedFiles;
+
+  final String? lcovGen;
+
   Config(
       {required this.percentage,
       required this.excludeSuffixes,
+      required this.includeUntestedFiles,
       required this.log,
       this.packageName,
-      this.command});
+      this.command,
+      this.lcovGen});
 }