+++ lib/models/config_model.dart
@@ -16,10 +16,17 @@ class ConfigModel {
   /// Command
   final ArgResults? command;
 
+  // If should include untested files to report
+  final bool includeUntestedFiles;
+
+  final String? lcovGen;
+
   ConfigModel(
       {required this.percentage,
       required this.excludeSuffixes,
+      required this.includeUntestedFiles,
       required this.log,
       this.packageName,
-      this.command});
+      this.command,
+      this.lcovGen});
 }